+++ dvc.lock
@@ -37,21 +37,35 @@ stages:
       md5: b124c02b39328021828fe8a17727d0c9
       size: 2565327
   train_model:
-    cmd: python src/models/train_model.py
+    cmd:
+    - python src/models/train_base_model.py
+    - python src/models/train_tuned_model.py
     deps:
     - path: data/train.csv
       hash: md5
       md5: b124c02b39328021828fe8a17727d0c9
       size: 2565327
-    - path: src/models/train_model.py
+    - path: src/models/train_base_model.py
       hash: md5
-      md5: e3e3c1e0fcf5896acb0bd72f182ad614
-      size: 1144
+      md5: 72d2a43c41919e2cfb5406b17f00672f
+      size: 581
+    - path: src/models/train_tuned_model.py
+      hash: md5
+      md5: 33e7fbd0980bd199d43c348a8093f28a
+      size: 723
+    - path: src/models/utils.py
+      hash: md5
+      md5: e1ab9525277942635929311062d8e43b
+      size: 664
     outs:
-    - path: models/model.pkl
+    - path: models/base_rf_model.pkl
       hash: md5
-      md5: b9d39c597fcea9393ec7777fd017daf9
-      size: 34199068
+      md5: 470f6dcc9ee768555ce8253520bfea88
+      size: 30782425
+    - path: models/tuned_rf_model.pkl
+      hash: md5
+      md5: 27c8a4b7b63fdfaf6949a4d9cdbd0a69
+      size: 703238687
   predict_model:
     cmd: python src/models/predict_model.py
     deps:
@@ -76,3 +90,40 @@ stages:
       hash: md5
       md5: 5bd49e3e36e33ae371bc1b67744c363c
       size: 155
+  train_base_model:
+    cmd:
+    - python src/models/train_base_model.py
+    deps:
+    - path: data/train.csv
+      hash: md5
+      md5: b124c02b39328021828fe8a17727d0c9
+      size: 2565327
+    - path: src/models/train_base_model.py
+      hash: md5
+      md5: 72d2a43c41919e2cfb5406b17f00672f
+      size: 581
+    - path: src/models/utils.py
+      hash: md5
+      md5: e1ab9525277942635929311062d8e43b
+      size: 664
+    outs:
+    - path: models/base_rf_model.pkl
+      hash: md5
+      md5: 470f6dcc9ee768555ce8253520bfea88
+      size: 30782425
+  train_tuned_model:
+    cmd:
+    - python src/models/train_tuned_model.py
+    deps:
+    - path: data/train.csv
+      hash: md5
+      md5: b124c02b39328021828fe8a17727d0c9
+      size: 2565327
+    - path: src/models/train_tuned_model.py
+      hash: md5
+      md5: 33e7fbd0980bd199d43c348a8093f28a
+      size: 723
+    - path: src/models/utils.py
+      hash: md5
+      md5: e1ab9525277942635929311062d8e43b
+      size: 664
