+++ dvc.yaml
@@ -2,37 +2,66 @@ stages:
   make_dataset:
     cmd: python src/data/make_dataset.py
     deps:
-    - data/urls_with_features.csv
-    - src/data/make_dataset.py
+      - data/urls_with_features.csv
+      - src/data/make_dataset.py
     outs:
-    - data/urls_with_features_selected.csv:
-        cache: false
+      - data/urls_with_features_selected.csv:
+          cache: false
   build_features:
     cmd: python src/features/build_features.py
     deps:
-    - data/urls_with_features_selected.csv
-    - src/features/build_features.py
+      - data/urls_with_features_selected.csv
+      - src/features/build_features.py
     outs:
-    - data/test.csv:
-        cache: false
-    - data/train.csv:
-        cache: false
-  train_model:
-    cmd: python src/models/train_model.py
+      - data/test.csv:
+          cache: false
+      - data/train.csv:
+          cache: false
+  train_base_model:
+    cmd:
+      - python src/models/train_base_model.py
     deps:
-    - data/train.csv
-    - src/models/train_model.py
+      - src/models/utils.py
+      - data/train.csv
+      - src/models/train_base_model.py
     outs:
-    - models/model.pkl:
-        cache: false
-  predict_model:
-    cmd: python src/models/predict_model.py
+      - models/base_rf_model.pkl:
+          cache: false
+
+  train_tuned_model:
+    cmd:
+      - python src/models/train_tuned_model.py
+    
     deps:
-    - models/model.pkl
-    - data/test.csv
-    - src/models/predict_model.py
+      - src/models/utils.py
+      - data/train.csv
+      - src/models/train_tuned_model.py
+
+  predict_base_model:
+    cmd:
+      - python src/models/predict_base_model.py
+    deps:
+      - src/models/utils.py
+      - models/base_rf_model.pkl
+      - data/test.csv
+      - src/models/predict_base_model.py
     outs:
-    - reports/classification_report.json:
-        cache: false
-    - reports/confusion_matrix.json:
-        cache: false
+      - reports/classification_report_base_rf.json:
+          cache: false
+      - reports/confusion_matrix_base_rf.json:
+          cache: false
+  
+  predict_tuned_model:
+    cmd:
+      - python src/models/predict_tuned_model.py
+    deps:
+      - src/models/utils.py
+      - models/tuned_rf_model.pkl
+      - data/test.csv
+      - src/models/predict_tuned_model.py
+    outs:
+      - reports/classification_report_tuned_rf.json:
+          cache: false
+      - reports/confusion_matrix_tuned_rf.json:
+          cache: false
+
